+++ gui/Pages/SolutionDomainView.py
@@ -619,7 +619,7 @@ class MeshInputDialog(QFileDialog):
         caption =  self.tr("Select input mesh file or directory")
         self.setWindowTitle(caption)
 
-        self.name_filter = str(self.tr("Imported or preprocessed meshes (mesh_input mesh_output)"))
+        self.name_filter = str(self.tr("Imported or preprocessed meshes (mesh_input mesh_output mesh_input_folder *.csm)"))
         self.setNameFilter(self.name_filter)
 
         self.select_label = str(self.tr("Select"))